backend/requirements.txt
boto3==1.34.29
Flask==2.2.2
Flask_RESTful==0.3.10
marshmallow==3.20.2
pandas==2.0.3
requests
gunicorn==19.4.5
python-dotenv==1.0.0